Store compile mode as enum in BuildConfig
