+++ src/cargo/ops/cargo_clean.rs
@@ -1,10 +1,10 @@
 use std::fs;
 use std::path::Path;
 
-use core::compiler::{BuildConfig, BuildContext, Context, Kind, Unit};
+use core::compiler::{BuildConfig, BuildContext, CompileMode, Context, Kind, Unit};
 use core::profiles::ProfileFor;
 use core::Workspace;
-use ops::{self, CompileMode};
+use ops;
 use util::errors::{CargoResult, CargoResultExt};
 use util::paths;
 use util::Config;
@@ -88,7 +88,8 @@ pub fn clean(ws: &Workspace, opts: &CleanOptions) -> CargoResult<()> {
         }
     }
 
-    let mut build_config = BuildConfig::new(config, Some(1), &opts.target, None)?;
+    let mut build_config =
+        BuildConfig::new(config, Some(1), &opts.target, None, CompileMode::Build)?;
     build_config.release = opts.release;
     let bcx = BuildContext::new(
         ws,
