+++ src/cargo/ops/cargo_package.rs
@@ -10,7 +10,7 @@ use git2;
 use tar::{Archive, Builder, EntryType, Header};
 
 use core::{Package, Source, SourceId, Workspace};
-use core::compiler::{DefaultExecutor, MessageFormat};
+use core::compiler::{CompileMode, DefaultExecutor, MessageFormat};
 use sources::PathSource;
 use util::{self, internal, Config, FileLock};
 use util::paths;
@@ -348,7 +348,7 @@ fn run_verify(ws: &Workspace, tar: &FileLock, opts: &PackageOpts) -> CargoResult
             },
             release: false,
             message_format: MessageFormat::Human,
-            mode: ops::CompileMode::Build,
+            mode: CompileMode::Build,
             target_rustdoc_args: None,
             target_rustc_args: None,
             export_dir: None,
